Basic Component Test 쪽 코드는 Binder 적용완료.

+++ Toy/UserInterface/TextureSourceBinder/TextureSourceInfo.h
@@ -14,7 +14,7 @@ struct TextureSourceInfo : public TextureInfo
 
 	bool LoadResource(ITextureLoad* load);
 	void SerializeIO(JsonOperation& operation);
-	inline Rectangle GetSource(int index) const noexcept { return (sources.size() > index) ? sources.at(index) : Rectangle{}; }
+	inline Rectangle GetSource(size_t index) const noexcept { return (sources.size() > index) ? sources.at(index) : Rectangle{}; }
 
 	using TextureInfo::GetIndex;
